feat: field.Hash
