+++ ecc/bn254/hash_to_g2.go
@@ -156,7 +156,7 @@ func MapToG2(u fptower.E2) G2Affine {
 func EncodeToG2(msg, dst []byte) (G2Affine, error) {
 
 	var res G2Affine
-	u, err := hashToFp(msg, dst, 2)
+	u, err := fp.Hash(msg, dst, 2)
 	if err != nil {
 		return res, err
 	}
@@ -175,7 +175,7 @@ func EncodeToG2(msg, dst []byte) (G2Affine, error) {
 // dst stands for "domain separation tag", a string unique to the construction using the hash function
 // https://www.ietf.org/archive/id/draft-irtf-cfrg-hash-to-curve-16.html#roadmap
 func HashToG2(msg, dst []byte) (G2Affine, error) {
-	u, err := hashToFp(msg, dst, 2*2)
+	u, err := fp.Hash(msg, dst, 2*2)
 	if err != nil {
 		return G2Affine{}, err
 	}
