+++ ecc/bw6-761/hash_to_g1.go
@@ -219,27 +219,6 @@ func g1EvalPolynomial(z *fp.Element, monic bool, coefficients []fp.Element, x *f
 	z.Set(&dst)
 }
 
-// hashToFp hashes msg to count prime field elements.
-// https://tools.ietf.org/html/draft-irtf-cfrg-hash-to-curve-06#section-5.2
-func hashToFp(msg, dst []byte, count int) ([]fp.Element, error) {
-	// 128 bits of security
-	// L = ceil((ceil(log2(p)) + k) / 8), where k is the security parameter = 128
-	const Bytes = 1 + (fp.Bits-1)/8
-	const L = 16 + Bytes
-
-	lenInBytes := count * L
-	pseudoRandomBytes, err := ecc.ExpandMsgXmd(msg, dst, lenInBytes)
-	if err != nil {
-		return nil, err
-	}
-
-	res := make([]fp.Element, count)
-	for i := 0; i < count; i++ {
-		res[i].SetBytes(pseudoRandomBytes[i*L : (i+1)*L])
-	}
-	return res, nil
-}
-
 // g1Sgn0 is an algebraic substitute for the notion of sign in ordered fields
 // Namely, every non-zero quadratic residue in a finite field of characteristic =/= 2 has exactly two square roots, one of each sign
 // https://www.ietf.org/archive/id/draft-irtf-cfrg-hash-to-curve-16.html#name-the-sgn0-function
@@ -269,7 +248,7 @@ func MapToG1(u fp.Element) G1Affine {
 func EncodeToG1(msg, dst []byte) (G1Affine, error) {
 
 	var res G1Affine
-	u, err := hashToFp(msg, dst, 1)
+	u, err := fp.Hash(msg, dst, 1)
 	if err != nil {
 		return res, err
 	}
@@ -287,7 +266,7 @@ func EncodeToG1(msg, dst []byte) (G1Affine, error) {
 // dst stands for "domain separation tag", a string unique to the construction using the hash function
 // https://www.ietf.org/archive/id/draft-irtf-cfrg-hash-to-curve-16.html#roadmap
 func HashToG1(msg, dst []byte) (G1Affine, error) {
-	u, err := hashToFp(msg, dst, 2*1)
+	u, err := fp.Hash(msg, dst, 2*1)
 	if err != nil {
 		return G1Affine{}, err
 	}
